requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.6.0
opencv-contrib-python>=4.6.0

# Deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Tracking
filterpy>=1.4.5
scikit-learn>=1.0.0
scipy>=1.7.0

# Computer vision utilities
Pillow>=9.0.0
matplotlib>=3.5.0

# Data handling
pandas>=1.3.0
pyyaml>=6.0

# API (optional)
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.28.0

# Utilities
tqdm>=4.62.0
colorama>=0.4.4

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Development
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
